+++ src/api/index.ts
@@ -3,7 +3,7 @@ import { getAuth } from 'firebase/auth';
 import { getFirestore } from 'firebase/firestore';
 import { getStorage } from 'firebase/storage';
 
-import { firebaseConfig } from '../constants/api';
+import { firebaseConfig } from '@/constants/api';
 
 const app = initializeApp(firebaseConfig);
