+++ src/components/common/buttons/Button.tsx
@@ -1,4 +1,4 @@
-import { ButtonHTMLAttributes, ReactNode } from 'react';
+import { ReactNode } from 'react';
 
 import { css, SerializedStyles } from '@emotion/react';
