+++ src/main.tsx
@@ -4,10 +4,10 @@ import { ThemeProvider } from '@emotion/react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 
-import App from './App';
-import GlobalStyle from './styles/GlobalStyle';
-import theme from './styles/theme';
+import App from '@/App';
 import { store } from '@/store/store';
+import GlobalStyle from '@/styles/GlobalStyle';
+import theme from '@/styles/theme';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
