+++ src/types/schedule.d.ts
