+++ src/utils/validators.ts
@@ -1,3 +1,3 @@
-import { ValidProps } from '../types/props';
+import { ValidProps } from '@/types/props';
 
 export const inputValid = ({ value, regex }: ValidProps) => regex.test(value.trim());
